Add browser build for ml-header-gen
Update config and scripts to create a browser build.
Move cli specific features from main.ts to cli.ts.
Use separate tsconfig for cli.ts compile.
